refactor: betAmount를 Wrapper 타입에서 스스로 검증하도록 수정
