+++ src/main/java/domain/participant/Name.java
@@ -3,10 +3,10 @@ package domain.participant;
 public record Name(
         String name
 ) {
-    private static final String PLAYER_NAME_LENGTH_ERROR = "플레이어의 이름은 2글자 이상 10글자 이하여야 합니다.";
-
     public static final int MINIMUM_BOUND = 2;
     public static final int MAXIMUM_BOUND = 10;
+
+    private static final String PLAYER_NAME_LENGTH_ERROR = "플레이어의 이름은 2글자 이상 10글자 이하여야 합니다.";
 
     public Name {
         validateNameLength(name);
